refactor: 예약 추가 요청 DTO 이름 변경
