+++ src/main/java/roomescape/dto/CreateReservationRequest.java
@@ -1,9 +1,10 @@
-package roomescape;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.Reservation;
 
-public record ReservationRequest(String name, LocalDate date, LocalTime time) {
+public record CreateReservationRequest(String name, LocalDate date, LocalTime time) {
 
     public Reservation toReservation(final long id) {
         return new Reservation(id, name, date, time);
