+++ src/main/java/roomescape/dto/CreateTimeSlotRequest.java
@@ -1,6 +1,7 @@
-package roomescape;
+package roomescape.dto;
 
 import java.time.LocalTime;
+import roomescape.ReservationTimeSlot;
 
 public record CreateTimeSlotRequest(
     LocalTime startAt
