+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -3,13 +3,13 @@ package roomescape.repository;
 import java.util.List;
 import java.util.Optional;
 import roomescape.Reservation;
-import roomescape.ReservationRequest;
+import roomescape.dto.CreateReservationRequest;
 
 public interface ReservationRepository {
 
     Optional<Reservation> findById(long id);
 
-    long save(ReservationRequest request);
+    long save(CreateReservationRequest request);
 
     boolean removeById(long id);
